requirements.txt
django
djangorestframework
pdfkit
qrcode